chore: removes extra avatar examples
