--- src/avatar/examples-list.js
@@ -1,14 +1,0 @@
-/*
-Copyright (c) 2018 Uber Technologies, Inc.
-
-This source code is licensed under the MIT license found in the
-LICENSE file in the root directory of this source tree.
-*/
-/* eslint-env node */
-/* eslint-disable flowtype/require-valid-file-annotation */
-
-module.exports = {
-  AVATAR_EXAMPLE: 'Avatar',
-  AVATAR_ERROR: 'Avatar error state',
-  AVATAR_OVERRIDE: 'Avatar override',
-};
